Double precision for lon lat (#2693)
* Introduce Point2d and Vector2d

* Use double precision for GeoPoint

* Read ll from json as double

* Fix unit tests

* Fix double precision for height action

* Continue fix tests

* hopefully reduce pointll hash collisions, fix long standing but benign bugs in decode7/encode7

* allow differnt precision on encode/decode7

* use 7 digits of precision for finding other places precision is lost

* make 7digit precision possible with a preprocessor defintion

* store 7bits of precision in OSMNode daat tile building time

* Fix hash collisions for PointXY

* Fix tiles unit test

* Don't loose precision while setting lon,lat to OSMNode

* Fix test_trimming

* Fix thor_worker test

* Fix matrix test

* 7 digits of precision for nodeinfo ll

* refactor TileHierarchy while tracking down a typo :(

* update isochrone test

* update summary shape

* fix map match test

* fix astar test

* lint

* update summary

* use double precision for path edges in loki search

* use our own hash_combine since older boost doesnt ahve it at same place

* more doubles in place of floats

* more doubles to shut up clang

* working on bss tests

* fixed one test in the suite

* slight differences to bss testing

* roll our own hash_combine for since boost relocates it

* forgot to remove header

* explicit instantiations

* tidy

* smarter isochrone test

* cleanup comments etc

* changelog derp

+++ valhalla/thor/triplegbuilder.h
@@ -24,7 +24,7 @@ namespace thor {
 struct EdgeTrimmingInfo {
   bool trim;
   midgard::PointLL vertex;
-  float distance_along;
+  double distance_along;
 };
 
 /**
